Cleared struct of project
