+++ CMakeLists.txt
@@ -2,4 +2,6 @@ CMAKE_MINIMUM_REQUIRED(VERSION 3.10)
 
 PROJECT(MoonDB C CXX)
 
-ADD_EXECUTABLE(MoonDB main.cpp)
+INCLUDE_DIRECTORIES(Include)
+
+ADD_EXECUTABLE(MoonDB Source/Main.cpp)
