+++ Include/CowConfig.hpp
