+++ Source/Main.cpp
@@ -1,13 +1,23 @@
 #include <iostream>
 #include <vector>
 #include <string>
-#include "MooDb/MooDb.hpp"
+#include "MooDb.hpp"
 
 template <class T>
 void PrintVector(std::vector< T > vec){
 	std::cout << std::endl;
 	for (int i = 0; i < vec.size(); i++){
 		std::cout << vec[i];
+		if (i != vec.size() - 1)
+			std::cout << ", ";
+	}
+}
+
+template <class T>
+void PrintVector(std::vector< T* > vec){
+	std::cout << std::endl;
+	for (int i = 0; i < vec.size(); i++){
+		std::cout << *vec[i];
 		if (i != vec.size() - 1)
 			std::cout << ", ";
 	}
@@ -27,14 +37,16 @@ int main() {
 	hp.Print();
 	hp.Save("Medications.csv");
 
-	std::vector< std::vector < std::string > > r = hp.GetRowsWithValue("RequiresID", "True");
+	std::vector< std::vector < std::string >* > r = hp.GetRowsWithValue("RequiresID", "True");
 	for (int i = 0; i < r.size(); i++){
-		PrintVector(r[i]);
+		PrintVector(*r[i]);
 		std::cout << std::endl;
 	}
 
 	Table tb = Table::Load("Patients.csv");
 	tb.Print();
-	PrintVector(tb.GetColumn("Name"));
-	PrintVector(tb.GetRow(0));
+	*tb.GetColumn("Name").at(1) = "REDACTED";
+	PrintVector<std::string>(tb.GetColumn("Name"));
+	tb.GetRow(0)->at(0) = "REDACTED"; // sets the first value in the first row
+	PrintVector(*tb.GetRow(0));
 }
